refactor(receipts): split schema and validation helpers
